chore: refactored constructors & naming
