--- src/__mocks__/axios.ts
@@ -1,7 +1,0 @@
-import type { AxiosStatic } from 'axios';
-
-const mockAxios = jest.createMockFromModule<AxiosStatic>('axios');
-
-mockAxios.create = jest.fn(() => mockAxios);
-
-export default mockAxios;
