+++ src/main/app/document/DocumentManagementClient.ts
@@ -1,19 +1,17 @@
 import Axios, { AxiosInstance, AxiosResponse } from 'axios';
 import FormData from 'form-data';
 
-import type { UserDetails } from '../../app/controller/AppRequest';
+import type { UserDetails } from '../controller/AppRequest';
 
 export class DocumentManagementClient {
   client: AxiosInstance;
-  userId: string;
 
-  constructor(options: Options) {
-    this.userId = options.user.id;
+  constructor(baseURL: string, authToken: string, private readonly user: UserDetails) {
     this.client = Axios.create({
-      baseURL: options.url,
+      baseURL,
       headers: {
-        Authorization: `Bearer ${options.user.accessToken}`,
-        ServiceAuthorization: options.authToken,
+        Authorization: `Bearer ${user.accessToken}`,
+        ServiceAuthorization: authToken,
       },
     });
   }
@@ -33,20 +31,14 @@ export class DocumentManagementClient {
     }
 
     const response = await this.client.post('/documents', formData, {
-      headers: { ...formData.getHeaders(), 'user-id': this.userId },
+      headers: { ...formData.getHeaders(), 'user-id': this.user.id },
     });
     return response.data?._embedded?.documents || [];
   }
 
   async delete({ url }: { url: string }): Promise<AxiosResponse> {
-    return this.client.delete(url, { headers: { 'user-id': this.userId } });
+    return this.client.delete(url, { headers: { 'user-id': this.user.id } });
   }
-}
-
-export interface Options {
-  url: string;
-  authToken: string;
-  user: UserDetails;
 }
 
 export interface DocumentManagementFile {
